requirements.txt
Flask
Flask-SQLAlchemy
PyJWT
bcrypt
waitress
PyYAML
Flask-CORS
requests
impmagic
zpp_store
cryptography